backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
pinecone-client==5.0.1
sentence-transformers==2.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
